Added one method in all repos. Implemented CRUD methods in services. Added more data to db. Started games and movies html files. Wrote tests for BookMapper with JUnit 4.

+++ src/main/java/com/przygodzki/bgm_app/service/BookService.java
@@ -8,7 +8,7 @@ public interface BookService {
 
     List<BookTo> findAll();
 
-    BookTo findById(Long id);
+    BookTo findById(Integer id);
 
     List<BookTo> findByTitle(String title);
 
@@ -16,7 +16,7 @@ public interface BookService {
 
     // BookTo update(BookTo book??);
 
-    void delete(Long id);
+    void delete(Integer id);
 
     // public List<BookTo> findByCriteria(String title, String author, String description);
 }
